Structural Changes and DB
Moved methods to helper files in order to keep the controller 'lean.'
Also, began populating the db with images, running the flickr scrape
continuously in a brute force method for now. The page is set with a
meta refresh every 5 seconds, and scrapes 500 images each time, adding
any new "colors" to the system.